+++ smart-contracts/contracts/cl-vault/src/helpers/generic.rs
@@ -1,0 +1,99 @@
+use crate::ContractError;
+use cosmwasm_std::{Coin, Decimal, Fraction, Uint128};
+
+pub fn with_slippage(amount: Uint128, slippage: Decimal) -> Result<Uint128, ContractError> {
+    let slippage_multiplier = Decimal::one().checked_sub(slippage)?;
+
+    let adjusted_amount = amount.checked_multiply_ratio(
+        slippage_multiplier.numerator(),
+        slippage_multiplier.denominator(),
+    )?;
+
+    Ok(adjusted_amount)
+}
+
+pub fn round_up_to_nearest_multiple(amount: i64, multiple: i64) -> i64 {
+    let remainder = amount % multiple;
+    if remainder == 0 {
+        amount
+    } else if amount < 0 {
+        amount - remainder
+    } else {
+        amount + multiple - remainder
+    }
+}
+
+pub fn sort_tokens(tokens: Vec<Coin>) -> Vec<Coin> {
+    let mut sorted_tokens = tokens;
+    sorted_tokens.sort_by(|a, b| a.denom.cmp(&b.denom));
+    sorted_tokens
+}
+
+pub fn extract_attribute_value_by_ty_and_key(
+    events: &[cosmwasm_std::Event],
+    ty: &str,
+    key: &str,
+) -> Option<String> {
+    events
+        .iter()
+        .filter(|event| event.ty == ty)
+        .flat_map(|event| event.attributes.iter())
+        .find(|attr| attr.key == key)
+        .map(|attr| attr.value.clone())
+}
+
+#[cfg(test)]
+mod tests {
+
+    use cosmwasm_std::coin;
+
+    use super::*;
+
+    #[test]
+    fn test_round_up_to_nearest_multiple() {
+        assert_eq!(round_up_to_nearest_multiple(10, 5), 10);
+        assert_eq!(round_up_to_nearest_multiple(11, 5), 15);
+        assert_eq!(round_up_to_nearest_multiple(12, 5), 15);
+        assert_eq!(round_up_to_nearest_multiple(13, 5), 15);
+        assert_eq!(round_up_to_nearest_multiple(14, 5), 15);
+        assert_eq!(round_up_to_nearest_multiple(15, 5), 15);
+        assert_eq!(round_up_to_nearest_multiple(16, 5), 20);
+        assert_eq!(round_up_to_nearest_multiple(17, 5), 20);
+        assert_eq!(round_up_to_nearest_multiple(18, 5), 20);
+        assert_eq!(round_up_to_nearest_multiple(19, 5), 20);
+        assert_eq!(round_up_to_nearest_multiple(20, 5), 20);
+        // does it also work for negative inputs?
+        assert_eq!(round_up_to_nearest_multiple(-10, 5), -10);
+        assert_eq!(round_up_to_nearest_multiple(-11, 5), -10);
+        assert_eq!(round_up_to_nearest_multiple(-12, 5), -10);
+        assert_eq!(round_up_to_nearest_multiple(-13, 5), -10);
+        assert_eq!(round_up_to_nearest_multiple(-14, 5), -10);
+        assert_eq!(round_up_to_nearest_multiple(-15, 5), -15);
+        assert_eq!(round_up_to_nearest_multiple(-16, 5), -15);
+        assert_eq!(round_up_to_nearest_multiple(-17, 5), -15);
+        assert_eq!(round_up_to_nearest_multiple(-18, 5), -15);
+        assert_eq!(round_up_to_nearest_multiple(-19, 5), -15);
+        assert_eq!(round_up_to_nearest_multiple(-20, 5), -20);
+    }
+
+    #[test]
+    fn test_sort_tokens() {
+        let tokens = vec![
+            coin(100, "uatom"),
+            coin(200, "uosmo"),
+            coin(300, "uqsr"),
+            coin(400, "ueth"),
+        ];
+
+        let expected = vec![
+            coin(100, "uatom"),
+            coin(400, "ueth"),
+            coin(200, "uosmo"),
+            coin(300, "uqsr"),
+        ];
+
+        let sorted_tokens = sort_tokens(tokens);
+
+        assert_eq!(sorted_tokens, expected);
+    }
+}
