+++ smart-contracts/contracts/cl-vault/src/state.rs
@@ -1,5 +1,7 @@
-use crate::vault::range::SwapDirection;
-use crate::vault::{helpers::CoinList, merge::CurrentMergeWithdraw};
+use crate::{
+    helpers::coinlist::CoinList,
+    vault::{merge::CurrentMergeWithdraw, range::SwapDirection},
+};
 use cosmwasm_schema::cw_serde;
 use cosmwasm_std::{Addr, Decimal, Decimal256, Uint128};
 use cw_dex_router::operations::SwapOperationsListUnchecked;
