+++ smart-contracts/contracts/cl-vault/src/vault/admin.rs
@@ -1,4 +1,4 @@
-use crate::helpers::assert_admin;
+use crate::helpers::assert::assert_admin;
 use crate::math::tick::build_tick_exp_cache;
 use crate::state::{
     Metadata, VaultConfig, ADMIN_ADDRESS, DEX_ROUTER, METADATA, RANGE_ADMIN, VAULT_CONFIG,
