+++ smart-contracts/contracts/cl-vault/src/vault/any_deposit.rs
@@ -9,10 +9,12 @@ use cw_dex_router::operations::{SwapOperationBase, SwapOperationsListUnchecked};
 use osmosis_std::types::osmosis::poolmanager::v1beta1::MsgSwapExactAmountInResponse;
 use osmosis_std::types::osmosis::tokenfactory::v1beta1::MsgMint;
 
-use crate::helpers::{
+use crate::helpers::assert::must_pay_one_or_two;
+use crate::helpers::getters::{
     get_asset0_value, get_depositable_tokens, get_single_sided_deposit_0_to_1_swap_amount,
     get_single_sided_deposit_1_to_0_swap_amount, get_twap_price,
 };
+use crate::helpers::msgs::swap_msg;
 use crate::query::query_total_vault_token_supply;
 use crate::reply::Replies;
 use crate::state::{PoolConfig, CURRENT_SWAP_ANY_DEPOSIT};
@@ -20,14 +22,12 @@ use crate::vault::concentrated_liquidity::get_cl_pool_info;
 use crate::vault::range::SwapDirection;
 use crate::vault::swap::SwapParams;
 use crate::{
-    helpers::must_pay_one_or_two,
     query::query_total_assets,
     state::{POOL_CONFIG, SHARES, VAULT_DENOM},
     vault::concentrated_liquidity::get_position,
     ContractError,
 };
 
-use super::helpers::swap_msg;
 use super::swap::SwapCalculationResult;
 
 pub fn execute_any_deposit(
